Fix how url changes happen
Currently, when $state.go is called the new url isn't updated properly. If
query parameters exist on the url they're removed because $state.go is called
with only the toState name.

There's an option for $state.go to not update the url which I enabled on
the successful validation of a permission (since you're not really going to
a new URL anyways). This leaves query parameters intact.