Add redirectTo syntax similar to  ui-sref
It is now possible to write something like:

redirectTo : 'my-state({ parameter: "value"})'

This new feature benefits string, object and function redirection. The regular
expression used to parse the string is the same that the one found in ui-sref.
To parse the parameter I use angular's $parse service.

I added one unit test to test that parameters are available in the state after
the redirection.

Note: It seems that ui-router filters parameters that are not declared the url
part of the state declaration